Fix and Abstract NoGuavaImmutable recipes
