+++ src/main/java/org/openrewrite/java/migrate/guava/NoGuavaImmutableMapOf.java
@@ -15,6 +15,7 @@
  */
 package org.openrewrite.java.migrate.guava;
 
+import com.google.common.collect.ImmutableMap;
 import org.openrewrite.Preconditions;
 import org.openrewrite.ExecutionContext;
 import org.openrewrite.Recipe;
@@ -31,154 +32,17 @@ import java.time.Duration;
 import java.util.*;
 import java.util.stream.Collectors;
 
-public class NoGuavaImmutableMapOf extends Recipe {
-    private static final MethodMatcher IMMUTABLE_MAP_MATCHER = new MethodMatcher("com.google.common.collect.ImmutableMap of(..)");
+public class NoGuavaImmutableMapOf extends AbstractNoGuavaImmutableOf {
 
     @Override
-    public String getDisplayName() {
-        return "Prefer `Map.of(..)` in Java 9 or higher";
+    String getGuavaType() {
+        return "com.google.common.collect.ImmutableMap";
     }
 
     @Override
-    public String getDescription() {
-        return "Replaces `ImmutableMap.of(..)` if the returned type is immediately down-cast.";
+    String getJavaType() {
+        return "java.util.Map";
     }
 
-    @Override
-    public Set<String> getTags() {
-        return new HashSet<>(Arrays.asList("RSPEC-4738", "guava"));
-    }
-
-    @Override
-    public Duration getEstimatedEffortPerOccurrence() {
-        return Duration.ofMinutes(10);
-    }
-
-    // Code is shared between `NoGuavaImmutableMapOf`, `NoGuavaImmutableListOf`, and `NoGuavaImmutableSetOf`.
-    // Updates to either may apply to each of the recipes.
-    @Override
-    public TreeVisitor<?, ExecutionContext> getVisitor() {
-        TreeVisitor<?, ExecutionContext> check = Preconditions.and(new UsesJavaVersion<>(9),
-                new UsesType<>("com.google.common.collect.ImmutableMap", false));
-        return Preconditions.check(check, new JavaIsoVisitor<ExecutionContext>() {
-
-            @Override
-            public J.MethodInvocation visitMethodInvocation(J.MethodInvocation method, ExecutionContext ctx) {
-                method = super.visitMethodInvocation(method, ctx);
-                if (IMMUTABLE_MAP_MATCHER.matches(method) && isParentTypeDownCast()) {
-                    maybeRemoveImport("com.google.common.collect.ImmutableMap");
-                    maybeAddImport("java.util.Map");
-
-                    String template = method.getArguments().stream()
-                            .map(arg -> {
-                                if (arg.getType() instanceof JavaType.Primitive) {
-                                    String type = "";
-                                    if (JavaType.Primitive.Boolean == arg.getType()) {
-                                        type = "Boolean";
-                                    } else if (JavaType.Primitive.Byte == arg.getType()) {
-                                        type = "Byte";
-                                    } else if (JavaType.Primitive.Char == arg.getType()) {
-                                        type = "Character";
-                                    } else if (JavaType.Primitive.Double == arg.getType()) {
-                                        type = "Double";
-                                    } else if (JavaType.Primitive.Float == arg.getType()) {
-                                        type = "Float";
-                                    } else if (JavaType.Primitive.Int == arg.getType()) {
-                                        type = "Integer";
-                                    } else if (JavaType.Primitive.Long == arg.getType()) {
-                                        type = "Long";
-                                    } else if (JavaType.Primitive.Short == arg.getType()) {
-                                        type = "Short";
-                                    } else if (JavaType.Primitive.String == arg.getType()) {
-                                        type = "String";
-                                    }
-                                    return TypeUtils.asFullyQualified(JavaType.buildType("java.lang." + type));
-                                } else {
-                                    return TypeUtils.asFullyQualified(arg.getType());
-                                }
-                            })
-                            .filter(Objects::nonNull)
-                            .map(type -> "#{any(" + type.getFullyQualifiedName() + ")}")
-                            .collect(Collectors.joining(",", "Map.of(", ")"));
-
-                    return JavaTemplate.builder(template)
-                            .contextSensitive()
-                            .imports("java.util.Map")
-                            .build()
-                            .apply(
-                                    updateCursor(method),
-                                    method.getCoordinates().replace(),
-                                    method.getArguments().get(0) instanceof J.Empty ? new Object[]{} : method.getArguments().toArray());
-                }
-                return method;
-            }
-
-            private boolean isParentTypeDownCast() {
-                J parent = getCursor().dropParentUntil(J.class::isInstance).getValue();
-                boolean isParentTypeDownCast = false;
-                if (parent instanceof J.VariableDeclarations.NamedVariable) {
-                    isParentTypeDownCast = isParentTypeMatched(((J.VariableDeclarations.NamedVariable) parent).getType());
-                } else if (parent instanceof J.Assignment) {
-                    J.Assignment a = (J.Assignment) parent;
-                    if (a.getVariable() instanceof J.Identifier && ((J.Identifier) a.getVariable()).getFieldType() != null) {
-                        isParentTypeDownCast = isParentTypeMatched(((J.Identifier) a.getVariable()).getFieldType().getType());
-                    } else if (a.getVariable() instanceof J.FieldAccess) {
-                        isParentTypeDownCast = isParentTypeMatched(a.getVariable().getType());
-                    }
-                } else if (parent instanceof J.Return) {
-                    // Does not currently support returns in lambda expressions.
-                    J j = getCursor().dropParentUntil(is -> is instanceof J.MethodDeclaration || is instanceof J.CompilationUnit).getValue();
-                    if (j instanceof J.MethodDeclaration) {
-                        TypeTree returnType = ((J.MethodDeclaration) j).getReturnTypeExpression();
-                        if (returnType != null) {
-                            isParentTypeDownCast = isParentTypeMatched(returnType.getType());
-                        }
-                    }
-                } else if (parent instanceof J.MethodInvocation) {
-                    J.MethodInvocation m = (J.MethodInvocation) parent;
-                    int index = 0;
-                    for (Expression argument : m.getArguments()) {
-                        if (IMMUTABLE_MAP_MATCHER.matches(argument)) {
-                            break;
-                        }
-                        index++;
-                    }
-                    if (m.getMethodType() != null) {
-                        isParentTypeDownCast = isParentTypeMatched(m.getMethodType().getParameterTypes().get(index));
-                    }
-                } else if (parent instanceof J.NewClass) {
-                    J.NewClass c = (J.NewClass) parent;
-                    int index = 0;
-                    if (c.getConstructorType() != null) {
-                        for (Expression argument : c.getArguments()) {
-                            if (IMMUTABLE_MAP_MATCHER.matches(argument)) {
-                                break;
-                            }
-                            index++;
-                        }
-                        if (c.getConstructorType() != null) {
-                            isParentTypeDownCast = isParentTypeMatched(c.getConstructorType().getParameterTypes().get(index));
-                        }
-                    }
-                } else if (parent instanceof J.NewArray) {
-                    J.NewArray a = (J.NewArray) parent;
-                    JavaType arrayType = a.getType();
-                    while (arrayType instanceof JavaType.Array) {
-                        arrayType = ((JavaType.Array) arrayType).getElemType();
-                    }
-
-                    isParentTypeDownCast = isParentTypeMatched(arrayType);
-                }
-                return isParentTypeDownCast;
-            }
-
-            private boolean isParentTypeMatched(@Nullable JavaType type) {
-                JavaType.FullyQualified fq = TypeUtils.asFullyQualified(type);
-                return TypeUtils.isOfClassType(fq, "java.util.Map")
-                       || TypeUtils.isOfClassType(fq, "java.lang.Object")
-                       || TypeUtils.isOfClassType(fq, "com.google.common.collect.ImmutableMap");
-            }
-        });
-    }
 }
